# backend/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/members": {
            "get": {
                "operationId": "MembersController_findAll",
                "summary": "Get info about all members",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MemberDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "members"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            },
            "post": {
                "operationId": "MembersController_create",
                "summary": "Create new member",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUpdateMemberDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MemberDTO"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "members"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
        "/api/members/{id}": {
            "get": {
                "operationId": "MembersController_findById",
                "summary": "Get full info about member",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MemberDTO"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "members"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            },
            "patch": {
                "operationId": "MembersController_update",
                "summary": "Update member",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUpdateMemberDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MemberDTO"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "members"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
        "/api/members/{id}/status": {
            "patch": {
                "operationId": "MembersController_freeze",
                "summary": "Freeze/unfreeze member",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateStatusDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MemberDTO"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "members"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
        "/api/members/{id}/telegram": {
            "patch": {
                "operationId": "MembersController_updateTelegramMetadata",
                "summary": "Add/update Telegram metadata for member",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTelegramMetadataDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Successful response"
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "members"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            },
            "delete": {
                "operationId": "MembersController_deleteTelegramMetadata",
                "summary": "Delete Telegram metadata for member",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful response"
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "members"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
        "/api/members/{id}/github": {
            "patch": {
                "operationId": "MembersController_updateGitHubMetadata",
                "summary": "Add/update GitHub metadata for member",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateGitHubMetadataDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Successful response"
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "members"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            },
            "delete": {
                "operationId": "MembersController_deleteGitHubMetadata",
                "summary": "Delete GitHub metadata for member",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful response"
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "members"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
        "/api/acs-keys/member/{memberId}": {
            "get": {
                "operationId": "ACSKeysController_findAllByMemberId",
                "summary": "Get ACS keys for specific member",
                "parameters": [
                    {
                        "name": "memberId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ACSKeyDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "acs-keys"
                ],
                "security": [
                    {
                        "cookie": []
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "RE: Swynca",
        "description": "B4CKSP4CE member/money management system",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "cookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "session"
            }
        },
        "schemas": {
...
            "MemberDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "frozen"
                        ]
                    },
                    "balance": {
                        "type": "string"
                    },
                    "joinedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "telegramMetadata": {
                        "$ref": "#/components/schemas/TelegramMetadataDTO"
                    },
                    "githubMetadata": {
                        "$ref": "#/components/schemas/GitHubMetadataDTO"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "email",
                    "status",
                    "balance",
                    "joinedAt"
                ]
            },
...
            "ErrorApiResponse": {
                "type": "object",
                "properties": {
                    "statusCode": {
                        "type": "number"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "required": [
                    "statusCode",
                    "message"
                ]
            },
...
            "CreateUpdateMemberDTO": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    }
                },
                "required": [
                    "name",
                    "email"
                ]
            },
            "UpdateStatusDTO": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "frozen"
                        ]
                    }
                },
                "required": [
                    "status"
                ]
            },
...
        }
    }
}